pyproject.toml
[project]
name = "beever-atlas"
version = "0.1.0"
description = "Wiki-first RAG system with dual semantic + graph memory"
readme = { text = "", content-type = "text/markdown" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "google-adk>=1.28.1",
    "litellm>=1.75.5",
    "weaviate-client>=4.10.0",
    "neo4j>=5.26.0",
    "pymongo>=4.10.0",
    "motor>=3.6.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.1.0",
    "pypdf>=6.10.2",
    "python-docx>=1.1.0",
    "cryptography>=42.0.0",
    "authlib>=1.6.11",
    "jellyfish>=1.0.0",
    "apscheduler>=4.0.0a5",
    "croniter>=1.3.0",
    "fastmcp>=3.2.2",
    "chardet>=5.2.0",
    "python-dateutil>=2.9.0",
    "langdetect>=1.0.9",
    "tavily-python>=0.5.0",
    "slowapi>=0.1.9",
    "aiolimiter>=1.1.0",
    "python-multipart>=0.0.26,<1.0.0",
]

[project.optional-dependencies]
nebula = [
    "nebula3-python>=3.4.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "hypothesis>=6.100.0",
    "markdown-it-py>=3.0.0",
    "pyright>=1.1.395",
]
test = [
    "tiktoken>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/beever_atlas"]

[tool.uv]
# Security pin: Mako <1.3.11 has a path-traversal vulnerability in
# TemplateLookup.get_template() (double-slash URI prefix). Mako is a
# transitive dep via alembic, so we constrain it here without adding a
# direct runtime import.
constraint-dependencies = [
    "mako>=1.3.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/__pycache__", ".venv", "dist", "build"]
typeCheckingMode = "basic"
reportMissingImports = "warning"
reportMissingTypeStubs = false
reportPrivateImportUsage = "none"
reportOptionalMemberAccess = "none"
reportGeneralTypeIssues = "warning"
reportArgumentType = "warning"
reportCallIssue = "warning"
reportOptionalSubscript = "warning"
reportAttributeAccessIssue = "warning"
reportReturnType = "warning"
reportAssignmentType = "warning"
reportOperatorIssue = "warning"
reportIndexIssue = "warning"
reportAbstractUsage = "warning"
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

[tool.coverage.run]
source = ["src/beever_atlas"]
omit = [
    "*/tests/*",
    "*/__main__.py",
]
branch = false

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@overload",
]
fail_under = 52
show_missing = false

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint.per-file-ignores]
# Re-export modules intentionally forward public API from submodules.
"src/beever_atlas/agents/consolidation/__init__.py" = ["F401"]
"src/beever_atlas/agents/media/__init__.py" = ["F401"]
"src/beever_atlas/llm/__init__.py" = ["F401"]
"src/beever_atlas/models/__init__.py" = ["F401", "F403"]
# Conditional imports gated on runtime env/feature flags; E402 is intentional.
"src/beever_atlas/server/app.py" = ["E402"]
"src/beever_atlas/stores/neo4j_store.py" = ["E402"]
"src/beever_atlas/api/channels.py" = ["E402"]
